+++ src/app/app/(with-app-bar)/layout.tsx
@@ -6,17 +6,22 @@ import { Search } from 'lucide-react';
 export default async function ApplicationLayout({ children }: React.PropsWithChildren) {
   return (
     <>
-      <main className='flex flex-col flex-1 px-default gap-2 py-2 overflow-y-scroll justify-stretch'>
+      <main className='flex flex-col flex-1 gap-2 overflow-y-scroll justify-stretch'>
         {children}
       </main>
-      <AppFooter>
-        <HighlightLink href='/app/feed'>
-          <button className='--ghost --round'>
-            <Search />
-          </button>
-        </HighlightLink>
-        <AddEventButton />
-      </AppFooter>
     </>
+  );
+}
+
+function Footer() {
+  return (
+    <AppFooter>
+      <HighlightLink href='/app/feed'>
+        <button className='--ghost --round'>
+          <Search />
+        </button>
+      </HighlightLink>
+      <AddEventButton />
+    </AppFooter>
   );
 }